Add a small margin for the auto refresh counter
If we wait until the maximum time before attempting to perform an auto refresh, then it is possible that a read/write command is pending at that time. This means that we wouldn't actually perform the refresh until *after* the required time. There is a chance that data could be corrupted.

By adding a small margin to the refresh counter, we can be sure that this will never occur. A read operation takes 5 clock cycles, so a margin of 10 will be sufficient.